+++ scripts/al1/user_data.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-set -ex
-
-yum update -y
-yum install -y openssh-clients