+++ CHANGELOG.md
@@ -5,6 +5,19 @@
 - [Amazon Linux 2 release notes](https://docs.aws.amazon.com/AL2/latest/relnotes/relnotes-al2.html)
 - [Amazon Linux AMI 2018.03 Release Notes](https://aws.amazon.com/amazon-linux-ami/2018.03-release-notes/)
 
+## 20251006
+- al2 ami version: 20251006
+- source al2 ami: amzn2-ami-minimal-hvm-2.0.20250929.2-x86_64-ebs
+- source al2 arm ami: amzn2-ami-minimal-hvm-2.0.20250929.2-arm64-ebs
+- source al2 kernel 5.10 ami:  amzn2-ami-minimal-hvm-2.0.20250929.2-x86_64-ebs
+- source al2 kernel 5.10 arm ami: amzn2-ami-minimal-hvm-2.0.20250929.2-arm64-ebs
+
+## 20251002
+- al2023 ami version: 20251002
+- source al2023 ami: al2023-ami-minimal-2023.9.20250929.0-kernel-6.1-x86_64
+- source al2023 arm ami: al2023-ami-minimal-2023.9.20250929.0-kernel-6.1-arm64
+- enhancement: Add ec2-instance-connect to AL2023 ECS AMIs [#544](https://github.com/aws/amazon-ecs-ami/pull/544)
+
 ## 20250923
 - ecs version: 1.99.1
 - al2 ami version: 20250923